feat(taxios-generate): programmatic usage
