+++ packages/taxios-generate/package.json
@@ -14,6 +14,8 @@
   "author": "Denis Karabaza <denis.karabaza@gmail.com>",
   "homepage": "https://github.com/simplesmiler/taxios/packages/taxios-generate",
   "license": "ISC",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "bin": "dist/bin.js",
   "repository": {
     "type": "git",
